requirements.txt
streamlit
opencv-python
numpy
Pillow
python-dotenv